refactor routes into modules; add/update save routes, config schema
