+++ backend/.sqlx/query-0368e5737412d5ae16c4319fa33289a9af1bf6e4768650baf82db7f236aeeeca.json
@@ -1,0 +1,100 @@
+{
+  "db_name": "PostgreSQL",
+  "query": "SELECT save.*,\n                      file.file_hash, file.file_filename,\n                      file.file_source_path, file.file_dest_path\n               FROM save\n               JOIN file USING(file_id)\n               WHERE save_id = ANY($1)",
+  "describe": {
+    "columns": [
+      {
+        "ordinal": 0,
+        "name": "save_id",
+        "type_info": "Uuid"
+      },
+      {
+        "ordinal": 1,
+        "name": "instance_id",
+        "type_info": "Uuid"
+      },
+      {
+        "ordinal": 2,
+        "name": "save_short_desc",
+        "type_info": "Varchar"
+      },
+      {
+        "ordinal": 3,
+        "name": "save_description",
+        "type_info": "Text"
+      },
+      {
+        "ordinal": 4,
+        "name": "file_id",
+        "type_info": "Uuid"
+      },
+      {
+        "ordinal": 5,
+        "name": "creator_id",
+        "type_info": "Uuid"
+      },
+      {
+        "ordinal": 6,
+        "name": "created_on",
+        "type_info": "Timestamptz"
+      },
+      {
+        "ordinal": 7,
+        "name": "state_derived_from",
+        "type_info": "Uuid"
+      },
+      {
+        "ordinal": 8,
+        "name": "save_derived_from",
+        "type_info": "Uuid"
+      },
+      {
+        "ordinal": 9,
+        "name": "replay_derived_from",
+        "type_info": "Uuid"
+      },
+      {
+        "ordinal": 10,
+        "name": "file_hash",
+        "type_info": "Bpchar"
+      },
+      {
+        "ordinal": 11,
+        "name": "file_filename",
+        "type_info": "Text"
+      },
+      {
+        "ordinal": 12,
+        "name": "file_source_path",
+        "type_info": "Text"
+      },
+      {
+        "ordinal": 13,
+        "name": "file_dest_path",
+        "type_info": "Text"
+      }
+    ],
+    "parameters": {
+      "Left": [
+        "UuidArray"
+      ]
+    },
+    "nullable": [
+      false,
+      false,
+      false,
+      false,
+      false,
+      false,
+      false,
+      true,
+      true,
+      true,
+      false,
+      false,
+      false,
+      false
+    ]
+  },
+  "hash": "0368e5737412d5ae16c4319fa33289a9af1bf6e4768650baf82db7f236aeeeca"
+}
