+++ frontend/frontend-web/index.html
@@ -7,7 +7,7 @@
         <title>Embedulator</title>
         <script src="v86/libv86.js"></script>
 <script id="config" type="application/json">
-    {"boot_into_record":false,"environment":{"created_on":"2025-04-15T22:52:00.553077Z","environment_config":null,"environment_core_name":"fceumm","environment_core_version":"1.52.0","environment_derived_from":null,"environment_framework":"retroarch","environment_id":"00000000-0000-0000-0000-000000000000","environment_name":"Nintendo Entertainment System"},"gisst_root":"https://localhost:3000","instance":{"created_on":"2025-04-15T22:52:01.509951Z","derived_from_instance":null,"derived_from_state":null,"environment_id":"00000000-0000-0000-0000-000000000000","instance_config":null,"instance_id":"00000000-0000-0000-0000-000000001001","work_id":"00000000-0000-0000-0000-000000001001"},"manifest":[{"file_dest_path":"2/c/e/b/a028e5747a6d0a658060644f56663d51-retroarch.cfg","file_filename":"retroarch.cfg","file_hash":"a028e5747a6d0a658060644f56663d51","file_source_path":"","object_id":"00000000-0000-0000-0000-000000000100","object_role":"config","object_role_index":0},{"file_dest_path":"a/9/4/9/1f21f648af663fbd3c864b7283d994c7-Alter Ego.nes","file_filename":"Alter Ego.nes","file_hash":"1f21f648af663fbd3c864b7283d994c7","file_source_path":"","object_id":"00000000-0000-0000-0000-000000001001","object_role":"content","object_role_index":0}],"save":null,"start":{"type":"cold"},"user":{"creator_id":"a7c734d0-e43f-4b18-b820-c90808f3f913","email":"test@test.edu","family_name":"familyname","given_name":"givenname","name":"test name","username":"test"},"work":{"created_on":"2025-04-15T22:52:01.458167Z","work_derived_from":null,"work_id":"00000000-0000-0000-0000-000000001001","work_name":"Alter Ego","work_platform":"Nintendo Entertainment System","work_version":"NTSC"}}
+    {"boot_into_record":false,"environment":{"created_on":"2025-04-15T22:52:00.553077Z","environment_config":null,"environment_core_name":"fceumm","environment_core_version":"1.52.0","environment_derived_from":null,"environment_framework":"retroarch","environment_id":"00000000-0000-0000-0000-000000000000","environment_name":"Nintendo Entertainment System"},"gisst_root":"https://localhost:3000","instance":{"created_on":"2025-04-15T22:52:01.509951Z","derived_from_instance":null,"derived_from_state":null,"environment_id":"00000000-0000-0000-0000-000000000000","instance_config":null,"instance_id":"00000000-0000-0000-0000-000000001001","work_id":"00000000-0000-0000-0000-000000001001"},"manifest":[{"file_dest_path":"2/c/e/b/a028e5747a6d0a658060644f56663d51-retroarch.cfg","file_filename":"retroarch.cfg","file_hash":"a028e5747a6d0a658060644f56663d51","file_source_path":"","object_id":"00000000-0000-0000-0000-000000000100","object_role":"config","object_role_index":0},{"file_dest_path":"a/9/4/9/1f21f648af663fbd3c864b7283d994c7-Alter Ego.nes","file_filename":"Alter Ego.nes","file_hash":"1f21f648af663fbd3c864b7283d994c7","file_source_path":"","object_id":"00000000-0000-0000-0000-000000001001","object_role":"content","object_role_index":0}],"save":null,"start":{"type":"cold"},"user":{"creator_id":"a7c734d0-e43f-4b18-b820-c90808f3f913","email":"test@test.edu","family_name":"familyname","given_name":"givenname","name":"test name","username":"test"},"work":{"created_on":"2025-04-15T22:52:01.458167Z","work_derived_from":null,"work_id":"00000000-0000-0000-0000-000000001001","work_name":"Alter Ego","work_platform":"Nintendo Entertainment System","work_version":"NTSC"},"saves":[]}
 </script>
   </head>
   <body>
